Make the honesty machinery apply to itself
Four things this project already believed, now enforced instead of remembered.

The translation coverage is pinned. tx() matches server English by exact string or
anchored pattern, and a miss falls through silently — English on a Spanish page,
no error, nothing to notice. It was complete only because it had been checked by
hand, string by string; the next reword would have undone that in silence, and had
already undone fourteen of them once, including the shift() look-ahead refusal, the
one message whose entire job is to stop a user shipping a lookahead bug. Now a
reworded string turns the suite red and names the file and line. The test reads the
tables out of i18n.js rather than keeping its own copy, and its docstring says what
it does NOT cover — a green test that quietly covers half the strings is the exact
failure this product is sold to prevent.

A crashed warm-up is no longer indistinguishable from a slow one. The exception sat
in an un-awaited task, `ready` stayed false forever, and /api/status kept answering
"warming up" — which is precisely the distinction that endpoint exists to make, so
it was failing at its own stated purpose. It now reports the failure, and
cd_deploy.sh stops on it instead of waiting out six minutes and rolling back with
no cause. The server half of that fix was inert until the deploy half landed.

cd_deploy.sh updates itself. It is installed by cd_setup.sh, so it was the one file
a deployment could not fix, and the drift only ever surfaced on the day a deploy
failed for a reason the newer version handled. After a deploy that worked it now
adopts the copy that came with it — after, never during: bash reads a script as it
runs it.

Lint is a gate. Twenty-five findings went to zero, so `continue-on-error` came off
and ruff is version-ranged: a minor bump ships new rules, and a gate that can go red
for something nobody pushed is a gate people learn to route around. Two of the
twenty-five were not style — a subprocess whose failure was ignored, and a
timezone-naive date in a codebase whose bar grids depend on UTC.

Also, for readers who are not programmers: the CSV importer no longer answers a bad
spreadsheet with a Python list repr or a pandas exception. It names the offending
value and the shapes it accepts. And a number now has one spelling per page —
the tile said 2,9% while the sentence explaining it said 2.9%, three lines apart.

+++ src/wavelab/validation/battery.py
@@ -236,7 +236,7 @@ def run_battery(
     # not forty. So we measure the mean run length of the signal itself.
     sig_flips = np.abs(np.diff(np.concatenate([[0.0], np.sign(sig)]))) > 0
     n_runs = max(1, int(sig_flips.sum()))
-    mean_dur = max(horizon_bars, int(round(active.sum() / n_runs)) if n_runs else horizon_bars)
+    mean_dur = max(horizon_bars, round(active.sum() / n_runs) if n_runs else horizon_bars)
     n_ef = effective_n(np.asarray(ts_ms)[: n][active], mean_dur, bar_ms)
     tests.append(Test(
         "effective_n", "Are there enough INDEPENDENT observations?",
